# docs/swagger.yaml
definitions:
...
  api.deleteBalitaRequest:
    properties:
      id:
        type: string
    type: object
  api.deleteBalitaResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
...
  object.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing stunting data
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stunting Web API
  version: 0.0.2
paths:
  /api/admin/balita/delete:
    delete:
      consumes:
      - application/json
      description: |-
        Soft delete balita data by setting deleted_date and deleted_id (Admin only)

        Performs soft delete operation:
        - Sets deleted_date to current timestamp
        - Sets deleted_id to current user ID
        - Data remains in database but is excluded from queries
        - Can be restored if needed in the future
        - Checks for related records before deletion
      parameters:
      - description: Balita ID to delete
        in: body
        name: balita
        required: true
        schema:
          $ref: '#/definitions/api.deleteBalitaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Balita deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.deleteBalitaResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Balita not found
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Delete balita data (soft delete)
      tags:
      - admin
...
  /api/admin/balita/restore:
    post:
      consumes:
      - application/json
      description: Restore soft deleted balita data by clearing deleted_date and deleted_id
        (Admin only)
      parameters:
      - description: Balita ID to restore
        in: body
        name: balita
        required: true
        schema:
          $ref: '#/definitions/api.deleteBalitaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Balita restored successfully
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.deleteBalitaResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Balita not found
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Restore deleted balita data
      tags:
      - admin
...
swagger: "2.0"